+++ satellite/src/main/java/com/linbit/linstor/core/apicallhandler/StltApiCallHandlerUtils.java
@@ -5,7 +5,6 @@ import com.linbit.linstor.annotation.ApiContext;
 import com.linbit.linstor.api.ApiCallRcImpl;
 import com.linbit.linstor.api.ApiConsts;
 import com.linbit.linstor.api.SpaceInfo;
-import com.linbit.linstor.core.ControllerPeerConnector;
 import com.linbit.linstor.core.CoreModule;
 import com.linbit.linstor.core.CoreModule.ExternalFileMap;
 import com.linbit.linstor.core.CoreModule.KeyValueStoreMap;
@@ -19,56 +18,39 @@ import com.linbit.linstor.core.CoreModule.StorPoolDefinitionMap;
 import com.linbit.linstor.core.DeviceManager;
 import com.linbit.linstor.core.StltExternalFileHandler;
 import com.linbit.linstor.core.apicallhandler.response.ApiRcException;
+import com.linbit.linstor.core.devmgr.StltReadOnlyInfo;
+import com.linbit.linstor.core.devmgr.StltReadOnlyInfo.ReadOnlyStorPool;
+import com.linbit.linstor.core.devmgr.StltReadOnlyInfo.ReadOnlyVlmProviderInfo;
 import com.linbit.linstor.core.identifier.ResourceName;
 import com.linbit.linstor.core.identifier.StorPoolName;
-import com.linbit.linstor.core.objects.Resource;
-import com.linbit.linstor.core.objects.StorPool;
 import com.linbit.linstor.core.objects.Volume;
-import com.linbit.linstor.dbdrivers.DatabaseException;
 import com.linbit.linstor.interfaces.StorPoolInfo;
 import com.linbit.linstor.layer.storage.DeviceProvider;
 import com.linbit.linstor.layer.storage.DeviceProviderMapper;
-import com.linbit.linstor.layer.storage.StorageLayer;
 import com.linbit.linstor.logging.ErrorReporter;
 import com.linbit.linstor.security.AccessContext;
 import com.linbit.linstor.security.AccessDeniedException;
 import com.linbit.linstor.storage.StorageException;
-import com.linbit.linstor.storage.interfaces.categories.resource.VlmProviderObject;
-import com.linbit.locks.LockGuard;
-import com.linbit.locks.LockGuardFactory;
-import com.linbit.locks.LockGuardFactory.LockObj;
-import com.linbit.locks.LockGuardFactory.LockType;
 import com.linbit.utils.Either;
 
+import javax.annotation.Nullable;
 import javax.inject.Inject;
-import javax.inject.Named;
 import javax.inject.Provider;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReadWriteLock;
 import java.util.function.Predicate;
-
-import static java.util.stream.Collectors.toList;
 
 public class StltApiCallHandlerUtils
 {
     private final ErrorReporter errorReporter;
     private final AccessContext apiCtx;
 
-    private final ReadWriteLock nodesMapLock;
-    private final ReadWriteLock storPoolDfnMapLock;
-    private final ControllerPeerConnector controllerPeerConnector;
-    private final ReadWriteLock rscDfnMapLock;
-    private final StorageLayer storageLayer;
     private final DeviceProviderMapper deviceProviderMapper;
-    private final LockGuardFactory lockGuardFactory;
     private final Provider<DeviceManager> devMgr;
     private final ExternalFileMap extFilesMap;
     private final StltExternalFileHandler stltExtFileHandler;
@@ -85,10 +67,6 @@ public class StltApiCallHandlerUtils
     public StltApiCallHandlerUtils(
         ErrorReporter errorReporterRef,
         @ApiContext AccessContext apiCtxRef,
-        ControllerPeerConnector controllerPeerConnectorRef,
-        @Named(CoreModule.NODES_MAP_LOCK) ReadWriteLock nodesMapLockRef,
-        @Named(CoreModule.STOR_POOL_DFN_MAP_LOCK) ReadWriteLock storPoolDfnMapLockRef,
-        @Named(CoreModule.RSC_DFN_MAP_LOCK) ReadWriteLock rscDfnMapLockRef,
         CoreModule.ExternalFileMap extFilesMapRef,
         StltExternalFileHandler stltExtFileHandlerRef,
         CoreModule.KeyValueStoreMap kvsMapRef,
@@ -99,18 +77,12 @@ public class StltApiCallHandlerUtils
         CoreModule.ResourceGroupMap rscGrpMapRef,
         CoreModule.ScheduleMap scheduleMapRef,
         CoreModule.StorPoolDefinitionMap storPoolDfnMapRef,
-        StorageLayer storageLayerRef,
         DeviceProviderMapper deviceProviderMapperRef,
-        LockGuardFactory lockGuardFactoryRef,
         Provider<DeviceManager> devMgrProviderRef
     )
     {
         errorReporter = errorReporterRef;
         apiCtx = apiCtxRef;
-        controllerPeerConnector = controllerPeerConnectorRef;
-        nodesMapLock = nodesMapLockRef;
-        storPoolDfnMapLock = storPoolDfnMapLockRef;
-        rscDfnMapLock = rscDfnMapLockRef;
         extFilesMap = extFilesMapRef;
         stltExtFileHandler = stltExtFileHandlerRef;
         kvsMap = kvsMapRef;
@@ -121,9 +93,7 @@ public class StltApiCallHandlerUtils
         rscGrpMap = rscGrpMapRef;
         scheduleMap = scheduleMapRef;
         storPoolDfnMap = storPoolDfnMapRef;
-        storageLayer = storageLayerRef;
         deviceProviderMapper = deviceProviderMapperRef;
-        lockGuardFactory = lockGuardFactoryRef;
         devMgr = devMgrProviderRef;
     }
 
@@ -134,136 +104,121 @@ public class StltApiCallHandlerUtils
     {
         Map<Volume.Key, Either<Long, ApiRcException>> allocatedMap = new HashMap<>();
 
-        try (LockGuard ignored = lockGuardFactory.build(
-            LockType.READ,
-            LockObj.NODES_MAP,
-            LockObj.RSC_DFN_MAP,
-            LockObj.STOR_POOL_DFN_MAP
-            )
-        )
+        StltReadOnlyInfo stltReadOnlyInfo = devMgr.get().getReadOnlyData();
+        // this method deliberately does not use the StorPoolInfo interface, since the getReadOnlyVolumes method is not
+        // part of that interface, since that would require StorPool also to implement it. For now this is not
+        // necessary, since this method should not be called / working with the actual StorPool and Volumes, but only
+        // with the read-only versions
+        Map<DeviceProvider, List<StltReadOnlyInfo.ReadOnlyStorPool>> storPoolsPerDeviceProvider = new HashMap<>();
+        for (ReadOnlyStorPool storPoolInfo : stltReadOnlyInfo.getStorPoolReadOnlyInfoList())
         {
-            Map<DeviceProvider, List<StorPool>> storPoolsPerDeviceProvider = new HashMap<>();
-            for (StorPool storPool : controllerPeerConnector.getLocalNode().streamStorPools(apiCtx).collect(toList()))
+            if (storPoolInfo.getDeviceProviderKind().usesThinProvisioning() &&
+                (storPoolFilter.isEmpty() || storPoolFilter.contains(storPoolInfo.getName())))
             {
-                if (storPool.getDeviceProviderKind().usesThinProvisioning() &&
-                    (storPoolFilter.isEmpty() || storPoolFilter.contains(storPool.getName())))
+                DeviceProvider deviceProvider = deviceProviderMapper.getDeviceProviderBy(storPoolInfo);
+                if (deviceProvider == null)
                 {
-                    DeviceProvider deviceProvider = deviceProviderMapper.getDeviceProviderBy(storPool);
-                    if (deviceProvider == null)
-                    {
-                        ApiRcException apiRcException = new ApiRcException(ApiCallRcImpl
+                    ApiRcException apiRcException = new ApiRcException(
+                        ApiCallRcImpl
                             .entryBuilder(
                                 ApiConsts.FAIL_UNKNOWN_ERROR,
-                                "Device provider for pool '" + storPool.getName() + "' not found"
+                                "Device provider for pool '" + storPoolInfo.getName() + "' not found"
                             )
                             .build()
-                        );
+                    );
 
-                        for (VlmProviderObject<Resource> vlmProviderObject : storPool.getVolumes(apiCtx))
-                        {
-                            Volume vlm = (Volume) vlmProviderObject.getVolume();
-                            if (
-                                resourceFilter.isEmpty() ||
-                                resourceFilter.contains(vlm.getResourceDefinition().getName())
-                            )
-                            {
-                                allocatedMap.put(vlm.getKey(), Either.right(apiRcException));
-                            }
-                        }
-                    }
-                    else
+                    for (ReadOnlyVlmProviderInfo readOnlyVlm : storPoolInfo.getReadOnlyVolumes())
                     {
-                        List<StorPool> list = storPoolsPerDeviceProvider.get(deviceProvider);
-                        if (list == null)
+                        if (resourceFilter.isEmpty() ||
+                            resourceFilter.contains(readOnlyVlm.getResourceName()))
                         {
-                            list = new ArrayList<>();
-                            storPoolsPerDeviceProvider.put(deviceProvider, list);
-                        }
-                        list.add(storPool);
-
-                    }
-                }
-            }
-
-            for (Entry<DeviceProvider, List<StorPool>> entry : storPoolsPerDeviceProvider.entrySet())
-            {
-                DeviceProvider deviceProvider = entry.getKey();
-
-                List<VlmProviderObject<Resource>> vlmDataList = new ArrayList<>();
-
-                for (StorPool storPool : entry.getValue())
-                {
-                    for (VlmProviderObject<Resource> vlmProviderObject : storPool.getVolumes(apiCtx))
-                    {
-                        Volume vlm = (Volume) vlmProviderObject.getVolume();
-                        if (resourceFilter.isEmpty() || resourceFilter.contains(vlm.getResourceDefinition().getName()))
-                        {
-                            vlmDataList.add(vlmProviderObject);
-                        }
-                    }
-                }
-
-                try
-                {
-                    deviceProvider.prepare(vlmDataList, Collections.emptyList());
-                    for (VlmProviderObject<Resource> vlmProviderObject : vlmDataList)
-                    {
-                        try
-                        {
-                            if (!vlmProviderObject.getRscLayerObject().hasIgnoreReason())
-                            {
-                                deviceProvider.updateAllocatedSize(vlmProviderObject);
-                                allocatedMap.put(
-                                    ((Volume) vlmProviderObject.getVolume()).getKey(),
-                                    Either.left(vlmProviderObject.getAllocatedSize())
-                                );
-                            }
-                        }
-                        catch (StorageException exc)
-                        {
-                            String reportNumber = errorReporter.reportError(exc);
-                            ApiRcException apiRcException = new ApiRcException(ApiCallRcImpl
-                                .entryBuilder(
-                                    ApiConsts.FAIL_UNKNOWN_ERROR,
-                                    "The device provider generated a StorageException. " +
-                                    "Error report number: " + reportNumber
-                                )
-                                .setCause(exc.getCauseText())
-                                .setDetails(exc.getDetailsText())
-                                .build()
-                            );
                             allocatedMap.put(
-                                ((Volume) vlmProviderObject.getVolume()).getKey(),
+                                readOnlyVlm.getVolumeKey(),
                                 Either.right(apiRcException)
                             );
                         }
                     }
                 }
-                catch (StorageException exc)
+                else
                 {
-                    String reportNumber = errorReporter.reportError(exc);
-                    ApiRcException apiRcException = new ApiRcException(ApiCallRcImpl
-                        .entryBuilder(
-                            ApiConsts.FAIL_UNKNOWN_ERROR,
-                            "The device provider generated a StorageException. Error report number: " + reportNumber
-                        )
-                        .setCause(exc.getCauseText())
-                        .setDetails(exc.getDetailsText())
-                        .build()
-                    );
-                    for (VlmProviderObject<Resource> vlmProviderObject : vlmDataList)
+                    List<StltReadOnlyInfo.ReadOnlyStorPool> list = storPoolsPerDeviceProvider.get(deviceProvider);
+                    if (list == null)
+                    {
+                        list = new ArrayList<>();
+                        storPoolsPerDeviceProvider.put(deviceProvider, list);
+                    }
+                    list.add(storPoolInfo);
+                }
+            }
+        }
+
+        for (Entry<DeviceProvider, List<StltReadOnlyInfo.ReadOnlyStorPool>> entry : storPoolsPerDeviceProvider
+            .entrySet())
+        {
+            DeviceProvider deviceProvider = entry.getKey();
+
+            List<ReadOnlyVlmProviderInfo> vlmDataList = new ArrayList<>();
+
+            for (StltReadOnlyInfo.ReadOnlyStorPool roStorPool : entry.getValue())
+            {
+                for (ReadOnlyVlmProviderInfo roVlm : roStorPool.getReadOnlyVolumes())
+                {
+                    if (resourceFilter.isEmpty() || resourceFilter.contains(roVlm.getResourceName()))
+                    {
+                        vlmDataList.add(roVlm);
+                    }
+                }
+            }
+
+            try
+            {
+                // we must not call deviceProvier.prepare(...), since that WILL interfere with the device-manager run.
+                // The prepare method populates provier-internal caches based on "lvs" or "zfs list". Recalculating
+                // those cached values and overriding based on our read-only-query while the device-manager is still
+                // relying on those cached values being based on the actual volumes that are currently being processed
+                // is a _very_ bad idea.
+
+                Map<ReadOnlyVlmProviderInfo, Long> allocatedSizeResult = deviceProvider.fetchAllocatedSizes(
+                    vlmDataList
+                );
+                for (Entry<ReadOnlyVlmProviderInfo, Long> allocatedSizeEntry : allocatedSizeResult.entrySet())
+                {
+                    @Nullable Long value = allocatedSizeEntry.getValue();
+                    if (value != null)
                     {
                         allocatedMap.put(
-                            ((Volume) vlmProviderObject.getVolume()).getKey(),
-                            Either.right(apiRcException)
+                            allocatedSizeEntry.getKey().getVolumeKey(),
+                            Either.left(value)
                         );
                     }
                 }
             }
-        }
-        catch (AccessDeniedException | DatabaseException exc)
-        {
-            errorReporter.reportError(new ImplementationError(exc));
+            catch (AccessDeniedException accDeniedExc)
+            {
+                throw new ImplementationError(accDeniedExc);
+            }
+            catch (StorageException exc)
+            {
+                String reportNumber = errorReporter.reportError(exc);
+                ApiRcException apiRcException = new ApiRcException(
+                    ApiCallRcImpl
+                        .entryBuilder(
+                            ApiConsts.FAIL_UNKNOWN_ERROR,
+                            "The device provider generated a StorageException. " +
+                                "Error report number: " + reportNumber
+                        )
+                        .setCause(exc.getCauseText())
+                        .setDetails(exc.getDetailsText())
+                        .build()
+                );
+                for (ReadOnlyVlmProviderInfo roVlm : vlmDataList)
+                {
+                    allocatedMap.put(
+                        roVlm.getVolumeKey(),
+                        Either.right(apiRcException)
+                    );
+                }
+            }
         }
 
         return allocatedMap;
@@ -285,30 +240,14 @@ public class StltApiCallHandlerUtils
     {
         Map<StorPoolInfo, Either<SpaceInfo, ApiRcException>> spaceMap = new HashMap<>();
 
-        Lock nodesMapReadLock = nodesMapLock.readLock();
-        Lock storPoolDfnMapReadLock = storPoolDfnMapLock.readLock();
+        StltReadOnlyInfo roInfo = devMgr.get().getReadOnlyData();
 
-        try
+        for (StorPoolInfo storPoolInfo : roInfo.getStorPoolReadOnlyInfoList())
         {
-            nodesMapReadLock.lock();
-            storPoolDfnMapReadLock.lock();
-
-            for (StorPool storPool : controllerPeerConnector.getLocalNode().streamStorPools(apiCtx).collect(toList()))
+            if (shouldIncludeSpTestRef.test(storPoolInfo))
             {
-                if (shouldIncludeSpTestRef.test(storPool))
-                {
-                    spaceMap.put(storPool, getStoragePoolSpaceInfoOrError(storPool));
-                }
+                spaceMap.put(storPoolInfo, getStoragePoolSpaceInfoOrError(storPoolInfo));
             }
-        }
-        catch (AccessDeniedException exc)
-        {
-            errorReporter.reportError(new ImplementationError(exc));
-        }
-        finally
-        {
-            storPoolDfnMapReadLock.unlock();
-            nodesMapReadLock.unlock();
         }
 
         return spaceMap;