+++ satellite/src/main/java/com/linbit/linstor/layer/storage/ebs/EbsInitiatorProvider.java
@@ -14,6 +14,7 @@ import com.linbit.linstor.core.CoreModule.RemoteMap;
 import com.linbit.linstor.core.StltConfigAccessor;
 import com.linbit.linstor.core.StltSecurityObjects;
 import com.linbit.linstor.core.apicallhandler.StltExtToolsChecker;
+import com.linbit.linstor.core.devmgr.StltReadOnlyInfo.ReadOnlyVlmProviderInfo;
 import com.linbit.linstor.core.objects.Resource;
 import com.linbit.linstor.core.objects.ResourceDefinition;
 import com.linbit.linstor.core.objects.ResourceGroup;
@@ -622,11 +623,17 @@ public class EbsInitiatorProvider extends AbsEbsProvider<LsBlkEntry>
         return pair == null ? null : pair.objB;
     }
 
-
     @Override
     protected void setDevicePath(EbsData<Resource> vlmDataRef, String devicePathRef) throws DatabaseException
     {
         vlmDataRef.setDevicePath(devicePathRef);
+    }
+
+    @Override
+    public Map<ReadOnlyVlmProviderInfo, Long> fetchAllocatedSizes(List<ReadOnlyVlmProviderInfo> vlmDataListRef)
+        throws StorageException
+    {
+        return fetchOrigAllocatedSizes(vlmDataListRef);
     }
 
     private static class TooManyDevicesException extends LinStorException