+++ satellite/src/main/java/com/linbit/linstor/layer/storage/exos/ExosProvider.java
@@ -10,6 +10,7 @@ import com.linbit.linstor.backupshipping.BackupShippingMgr;
 import com.linbit.linstor.clone.CloneService;
 import com.linbit.linstor.core.StltConfigAccessor;
 import com.linbit.linstor.core.apicallhandler.StltExtToolsChecker;
+import com.linbit.linstor.core.devmgr.StltReadOnlyInfo.ReadOnlyVlmProviderInfo;
 import com.linbit.linstor.core.identifier.ResourceName;
 import com.linbit.linstor.core.identifier.StorPoolName;
 import com.linbit.linstor.core.identifier.VolumeNumber;
@@ -1230,5 +1231,12 @@ public class ExosProvider extends AbsStorageProvider<ExosRestVolume, ExosData<Re
     protected long getExtentSize(AbsStorageVlmData<?> vlmDataRef)
     {
         return EXTENT_SIZE_IN_KIB;
+    }
+
+    @Override
+    public Map<ReadOnlyVlmProviderInfo, Long> fetchAllocatedSizes(List<ReadOnlyVlmProviderInfo> vlmDataListRef)
+        throws StorageException
+    {
+        return fetchOrigAllocatedSizes(vlmDataListRef);
     }
 }