+++ satellite/src/main/java/com/linbit/linstor/layer/storage/file/FileProvider.java
@@ -11,6 +11,7 @@ import com.linbit.linstor.clone.CloneService;
 import com.linbit.linstor.core.LinStor;
 import com.linbit.linstor.core.StltConfigAccessor;
 import com.linbit.linstor.core.apicallhandler.StltExtToolsChecker;
+import com.linbit.linstor.core.devmgr.StltReadOnlyInfo.ReadOnlyVlmProviderInfo;
 import com.linbit.linstor.core.identifier.ResourceName;
 import com.linbit.linstor.core.identifier.StorPoolName;
 import com.linbit.linstor.core.identifier.VolumeNumber;
@@ -849,5 +850,12 @@ public class FileProvider extends AbsStorageProvider<FileInfo, FileData<Resource
     protected String getStorageName(FileData<Resource> vlmDataRef) throws DatabaseException
     {
         return vlmDataRef.getStorageDirectory().toString();
+    }
+
+    @Override
+    public Map<ReadOnlyVlmProviderInfo, Long> fetchAllocatedSizes(List<ReadOnlyVlmProviderInfo> vlmDataListRef)
+        throws StorageException
+    {
+        return fetchOrigAllocatedSizes(vlmDataListRef);
     }
 }