+++ server/src/main/java/com/linbit/linstor/interfaces/NodeInfo.java
@@ -1,13 +1,5 @@
 package com.linbit.linstor.interfaces;
 
-/*
- * TODO this interface should be in the satellite project, not server.
- *
- * The only reason it is still in the server-project are the serializer-classes. For some reason satellite specific
- * methods (i.e. methods that only the satellite can send to the controller, not vice versa) are still in the server-
- * project instead of the satellite project.
- */
-
 import com.linbit.linstor.core.identifier.NodeName;
 import com.linbit.linstor.propscon.ReadOnlyProps;
 import com.linbit.linstor.security.AccessContext;
@@ -15,6 +7,13 @@ import com.linbit.linstor.security.AccessDeniedException;
 
 import java.util.UUID;
 
+/*
+ * TODO this interface should be in the satellite project, not server.
+ *
+ * The only reason it is still in the server-project are the serializer-classes. For some reason satellite specific
+ * methods (i.e. methods that only the satellite can send to the controller, not vice versa) are still in the server-
+ * project instead of the satellite project.
+ */
 public interface NodeInfo
 {
     UUID getUuid();