+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Improved responsiveness: Allow some API calls to be executed concurrently on the satellite
+
 ### Fixed
 
 - Snap, create: Fixed possible deadlock